feat: move api extraction to seperate file
